Busca de turma com maior quantidade aluno e a quantidade de aluno

+++ Backend/src/main/java/com/example/academia/core/entity/Turma.java
@@ -43,7 +43,7 @@ public class Turma implements Serializable{
 	private Professor professor;
 		
 	@Column(name = "qto_Aluno")
-	private int qtoAluno;
+	private Long qtoAluno;
 	
 	@Enumerated(EnumType.STRING)
     @Column(name = "status")
