+++ Backend/src/main/java/com/example/academia/v1/dto/TurmaRetornoDTO.java
@@ -19,7 +19,7 @@ public class TurmaRetornoDTO {
 	
 	private Aluno aluno;
 	
-	private int qtoAluno;
+	private Long qtoAluno;
 	
 	private int qtoProfessor;
